Function lal::build

pub fn build(
    cfg: &Config, 
    manifest: &Manifest, 
    opts: &BuildOptions, 
    envname: String, 
    _modes: ShellModes
) -> LalResult<()>

Runs the ./BUILD script in a container and packages artifacts.

The function performs basic sanity checks, before shelling out to docker run to perform the actual execution of the containerized ./BUILD script.