Function lal::verify

pub fn verify(m: &Manifest, env: &str, simple: bool) -> LalResult<()>

Verifies that ./INPUT satisfies all strictness conditions.

This first verifies that there are no key mismatches between defaultConfig and configurations in the manifest.

Once this is done, INPUT is analysed thoroughly via each components lockfiles. Missing dependencies, or multiple versions dependend on implicitly are both considered errors for verify, as are having custom versions in ./INPUT.

This function is meant to be a helper for when we want official builds, but also a way to tell developers that they are using things that differ from what jenkins would use.

A simple verify was added to aid the workflow of stashed components. Users can use lal verify --simple or lal build -s aka. --simple-verify, instead of having to use lal build --force when just using stashed components. This avoids problems with different environments going undetected.