Struct lal::Component

pub struct Component {
    pub name: String,
    pub version: u32,
    pub tarball: String,
}

The basic definition of a component as it exists online

A component may have many build artifacts from many environments.

Fields
name: String

Name of the component

version: u32

Version number

tarball: String

The raw URL of the tarball at the specified version number