Struct lal::ConfigDefaults

pub struct ConfigDefaults {
    pub backend: BackendConfiguration,
    pub environments: BTreeMap<String, Container>,
    pub mounts: Vec<Mount>,
    pub minimum_lal: Option<String>,
}

Representation of a configuration defaults file

This file is being used to generate the config when using lal configure

Fields
backend: BackendConfiguration

Configuration settings for the Backend

environments: BTreeMap<String, Container>

Environments shorthands that are allowed and their full meaning

mounts: Vec<Mount>

Extra volume mounts to be set for the container

minimum_lal: Option<String>

Optional minimum version restriction of lal

Methods

impl ConfigDefaults

fn read(file: &str) -> LalResult<ConfigDefaults>

Open and deserialize a defaults file