Struct lal::Manifest

pub struct Manifest {
    pub name: String,
    pub environment: String,
    pub components: BTreeMap<String, ComponentConfiguration>,
    pub dependencies: BTreeMap<String, u32>,
    pub devDependencies: BTreeMap<String, u32>,
    // some fields omitted
}

Representation of manifest.json

Fields
name: String

Name of the main component

environment: String

Default environment to build in

components: BTreeMap<String, ComponentConfiguration>

Components and their available configurations that are buildable

dependencies: BTreeMap<String, u32>

Dependencies that are always needed

devDependencies: BTreeMap<String, u32>

Development dependencies

Methods

impl Manifest

fn new(name: &str, env: &str, location: PathBuf) -> Manifest

Initialize a manifest struct based on a name

The name is assumed to be the default component and will create a component configuration for it with its default values.

fn all_dependencies(&self) -> BTreeMap<String, u32>

Merge dependencies and devDependencies into one convenience map

fn read() -> LalResult<Manifest>

Read a manifest file in PWD

fn read_from(pwd: &PathBuf) -> LalResult<Manifest>

Read a manifest file in an arbitrary path

fn write(&self) -> LalResult<()>

Update the manifest file in the current folder

fn verify(&self) -> LalResult<()>

Verify assumptions about configurations