Trait lal::CachedBackend

pub trait CachedBackend {
    fn retrieve_published_component(
        &self, 
        name: &str, 
        version: Option<u32>, 
        env: Option<&str>
    ) -> LalResult<(PathBuf, Component)>;

    fn retrieve_stashed_component(
        &self, 
        name: &str, 
        code: &str
    ) -> LalResult<PathBuf>;

    fn unpack_published_component(
        &self, 
        name: &str, 
        version: Option<u32>, 
        env: Option<&str>
    ) -> LalResult<Component>;

    fn unpack_stashed_component(&self, name: &str, code: &str) -> LalResult<()>;

    fn stash_output(&self, name: &str, code: &str) -> LalResult<()>;
}

A secondary trait that builds upon the Backend trait

This wraps the common fetch commands in a caching layer on the cache dir.

Required Methods

fn retrieve_published_component(
&self,
name: &str,
version: Option<u32>,
env: Option<&str>
) -> LalResult<(PathBuf, Component)>

Retrieve the location to a cached published component (downloading if necessary)

fn retrieve_stashed_component(
&self,
name: &str,
code: &str
) -> LalResult<PathBuf>

Retrieve the location to a stashed component

fn unpack_published_component(
&self,
name: &str,
version: Option<u32>,
env: Option<&str>
) -> LalResult<Component>

Retrieve and unpack a cached component in INPUT

fn unpack_stashed_component(&self, name: &str, code: &str) -> LalResult<()>

Retrieve and unpack a stashed component to INPUT

fn stash_output(&self, name: &str, code: &str) -> LalResult<()>

Add a stashed component from a folder